+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Unreleased
 ----------
 * Correct handling of l = u in IQP.
 * Correct handling of missing Ax=b in IQP.
+* Auto objective function scaling in IQP.
+* Basic IQP untitests.
 * Multistage stochastic solvers.
 
 Version 1.0